+++ packages/pk-cli/src/cli.ts
@@ -1,4 +1,4 @@
-import '@pkerschbaum/runtime-extensions-node';
+import '@patricktree/runtime-extensions-node';
 
 import * as commander from '@commander-js/extra-typings';
 import fs from 'node:fs';