+++ packages/pk-web-stack-codemods/src/codemods/01-monorepo-basis/codemod.mjs
@@ -6,7 +6,7 @@ import path from 'node:path';
 import url from 'node:url';
 import { $ } from 'zx';
 
-import { check } from '@pkerschbaum/commons-ecma/util/assert';
+import { check } from '@patricktree/commons-ecma/util/assert';
 
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));