+++ packages/pk-web-stack-codemods/src/pk-web-stack-codemods.mjs
@@ -5,7 +5,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import url from 'node:url';
 
-import { check } from '@pkerschbaum/commons-ecma/util/assert';
+import { check } from '@patricktree/commons-ecma/util/assert';
 
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));