+++ packages/codemod-rewrite-module-specifiers-to-full-paths/CHANGELOG.md
@@ -1,10 +1,10 @@
-# @pkerschbaum/codemod-rewrite-module-specifiers-to-full-paths
+# @patricktree/codemod-rewrite-module-specifiers-to-full-paths
 
 ## 1.0.1
 
 ### Patch Changes
 
-- 3f52b98: fix(codemod-rewrite-module-specifiers-to-full-paths): make `@pkerschbaum/runtime-extensions-node` a dev dependency
+- 3f52b98: fix(codemod-rewrite-module-specifiers-to-full-paths): make `@patricktree/runtime-extensions-node` a dev dependency
 
 ## 1.0.0